// server/types/Keys.enum.ts
export enum Keys {
  JWT_KEY_ACCESS = 'JWT_KEY_ACCESS',
  JWT_KEY_REFRESH = 'JWT_KEY_REFRESH',
  ACCESS_TOKEN_EXP = 'ACCESS_TOKEN_EXP',
  REFRESH_TOKEN_EXP = 'REFRESH_TOKEN_EXP',
  JWT_KEY_RECOVERY = 'JWT_KEY_RECOVERY',
  ACCESS_TOKEN = 'access_token',
  REFRESH_TOKEN = 'refresh_token',
  RECOVERY_TOKEN = 'recovery_token',
}